Fix Product URL In Breadcrumbs
The `storefront_url_for` method doesn't handle models other than taxons,
but the Schema.org helpers use it to render breadcrumb URLs in the
`BreadcrumbList` for any model that's in the breadcrumbs. To prevent
incorrect URLs from showing up in the breadcrumbs, Workarea defines a
new helper method called `breadcrumb_url_for` that will render the
`storefront_url_for` a taxon if the model passed in is a taxon,
otherwise it will discover the route name based on the demodulized class
of the object passed into breadcrumbs.

(#83)